requirements.txt
alembic==1.18.4
bandit
blinker==1.9.0
click==8.3.3
Deprecated==1.3.1
edgartools==5.30.0
email-validator==2.3.0
Flask==3.1.3
Flask-Admin==2.0.2
Flask-Bootstrap==3.3.7.1
Flask-Caching==2.4.0
Flask-Limiter==4.1.1
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.3.0
gunicorn==25.3.0
itsdangerous==2.2.0
Jinja2==3.1.6
limits==5.8.0
MarkupSafe==3.0.3
pillow==12.2.0
pyotp==2.9.0
pytest==9.0.3
pytest-flask==1.3.0
qrcode==8.2
requests==2.33.1
SQLAlchemy==2.0.49
Werkzeug==3.1.8
wrapt==2.1.2
WTForms==3.2.1
